Fix issue #13: Restructure README and reorganize agent files
- Moved agent.py, test_suite.py, strands_agent.py to agents/parking_lot/
- Added README.md in parking_lot explaining work-in-progress status
- Renamed agent_w_auth.py to agent.py
- Restructured README.md post-installation section:
  - Moved steps 3+ to new 'Using the Gateway and Registry with AI Agents' section
  - Added 'Run Agent with User Identity' subsection with .env.user setup
  - Added 'Run Agent with Its Own Agentic Identity' subsection with .env.agent setup
  - Documented JWT token flow and Cognito group authorization
