+++ agents/parking_lot/README.md
@@ -1,0 +1,3 @@
+# Parking Lot
+
+This code is work in progress and would be moved to the agents directory soon after appropriate modifications.
